# docs/swagger.yaml
basePath: /
definitions:
...
  model.Rutina:
    properties:
      fecha_creacion:
        type: string
      id:
        type: integer
      nombre:
        type: string
      objetivo:
        type: string
      publica:
        type: boolean
      usuario_id:
        type: integer
    type: object
...
  model.SesionEjercicio:
    properties:
      ejercicio_id:
        type: integer
      fecha:
        type: string
      id:
        type: integer
      observacion:
        type: string
      orden:
        type: integer
      peso:
        type: number
      repeticiones:
        type: integer
      series:
        type: integer
      sesion_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestión de rutinas de gimnasio
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GymBro API
  version: "1.0"
paths:
...
  /rutinas:
    get:
      consumes:
      - application/json
      description: Obtiene la lista completa de rutinas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Rutina'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener todas las rutinas
      tags:
      - rutinas
    post:
      consumes:
      - application/json
      description: Crea una nueva rutina en el sistema
      parameters:
      - description: Datos de la rutina
        in: body
        name: rutina
        required: true
        schema:
          $ref: '#/definitions/model.Rutina'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Rutina'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear nueva rutina
      tags:
      - rutinas
  /rutinas/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina una rutina del sistema
      parameters:
      - description: ID de la rutina
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Rutina eliminada correctamente
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar rutina
      tags:
      - rutinas
    get:
      consumes:
      - application/json
      description: Obtiene una rutina específica por su ID
      parameters:
      - description: ID de la rutina
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Rutina'
        "404":
          description: Rutina no encontrada
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener rutina por ID
      tags:
      - rutinas
    put:
      consumes:
      - application/json
      description: Actualiza una rutina existente
      parameters:
      - description: ID de la rutina
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados de la rutina
        in: body
        name: rutina
        required: true
        schema:
          $ref: '#/definitions/model.Rutina'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Rutina'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Rutina no encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar rutina
      tags:
      - rutinas
...
  /sesion-ejercicio:
    get:
      consumes:
      - application/json
      description: Obtiene la lista completa de ejercicios de sesión
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SesionEjercicio'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener todos los ejercicios de sesión
      tags:
      - sesion-ejercicio
    post:
      consumes:
      - application/json
      description: Crea un nuevo ejercicio de sesión en el sistema
      parameters:
      - description: Datos del ejercicio de sesión
        in: body
        name: sesion_ejercicio
        required: true
        schema:
          $ref: '#/definitions/model.SesionEjercicio'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.SesionEjercicio'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear nuevo ejercicio de sesión
      tags:
      - sesion-ejercicio
  /sesion-ejercicio/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un ejercicio de sesión del sistema
      parameters:
      - description: ID del ejercicio de sesión
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Ejercicio de sesión eliminado
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar ejercicio de sesión
      tags:
      - sesion-ejercicio
    get:
      consumes:
      - application/json
      description: Obtiene un ejercicio de sesión específico por su ID
      parameters:
      - description: ID del ejercicio de sesión
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SesionEjercicio'
        "404":
          description: Ejercicio de sesión no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener ejercicio de sesión por ID
      tags:
      - sesion-ejercicio
    put:
      consumes:
      - application/json
      description: Actualiza un ejercicio de sesión existente
      parameters:
      - description: ID del ejercicio de sesión
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados del ejercicio de sesión
        in: body
        name: sesion_ejercicio
        required: true
        schema:
          $ref: '#/definitions/model.SesionEjercicio'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SesionEjercicio'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar ejercicio de sesión
      tags:
      - sesion-ejercicio
  /sesion-ejercicio/sesion/{id}:
    get:
      consumes:
      - application/json
      description: Obtiene todos los ejercicios de una sesión específica
      parameters:
      - description: ID de la sesión
        in: path
        name: id
        required: true
        type: integer
      - description: 'Fecha desde (formato: 2006-01-02)'
        in: query
        name: fechaDesde
        type: string
      - description: 'Fecha hasta (formato: 2006-01-02)'
        in: query
        name: fechaHasta
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SesionEjercicio'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      summary: Obtener ejercicios de sesión por ID de sesión
      tags:
      - sesiones-ejercicios
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"